refactor: Refactored FlightSearchForm file
